+++ CHANGELOG.md
@@ -31,3 +31,4 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Fixed `BuyButton` minicart event
 - Fixed `item-quantity` product dispatch
 - Fixed font styling
+- Added vertical padding to price and inventory